[Refactor] 글 수정 시 이미지 수정여부

+++ src/main/java/com/travelmaker/service/PostServiceImpl.java
@@ -184,23 +184,28 @@ public class PostServiceImpl implements PostService {
                 // 수정할 게시글이 존재하지 않는 경우
                 .orElseThrow(() -> new CustomException(ErrorCode.NULL_VALUE)));
 
-        // TODO: 이미지가 수정되었는지 확인하는 값
-        // 저장된 이미지 삭제하기
-        String[] urls = entity.get().getPostImg().split(",");
-        for(String url : urls){
-            amazonS3ResourceStorage.deleteFile(url);
+        if(post.isEdited()) {   // 이미지가 수정되었을 때
+            // 저장된 이미지 삭제하기
+            String[] urls = entity.get().getPostImg().split(",");
+            for (String url : urls) {
+                amazonS3ResourceStorage.deleteFile(url);
+            }
+
+            // 이미지 재등록
+            String imageUrl = "";
+            for (int i = 0; i < images.size(); i++) {
+                FileDetail fileDetail = FileDetail.multipartOf(images.get(i));
+                String storedImg = amazonS3ResourceStorage.store(fileDetail.getPath(), images.get(i), fileDetail.getId());
+                // 쉼표(,)로 split
+                imageUrl += storedImg.split("/")[4] + ",";
+            }
+
+            post.setPostImg(imageUrl);
+        } else {
+            post.setPostImg(entity.get().getPostImg());
         }
 
-        // 이미지 재등록
-        String imageUrl = "";
-        for(int i = 0; i < images.size(); i++) {
-            FileDetail fileDetail = FileDetail.multipartOf(images.get(i));
-            String storedImg = amazonS3ResourceStorage.store(fileDetail.getPath(), images.get(i), fileDetail.getId());
-            // 쉼표(,)로 split
-            imageUrl += storedImg.split("/")[4] + ",";
-        }
-
-        post.setPostImg(imageUrl);
+        // TODO: 해시태그 수정
 
         Optional<Integer> updatedEntity = Optional.ofNullable(repository.updatePost(post.getIdx(), post.getTitle(), post.getContent(), post.getFigures().toString(), post.getPostImg())
                 // 수정 시 오류가 발생한 경우
